+++ tdd-samples/callback-infra/int/CMakeLists.txt
@@ -1,53 +1,10 @@
-project(callback-infra-int)
-# project(callback-infra-int C CXX)
-
-cmake_minimum_required(VERSION 2.8)
-
-enable_testing()
-
-####################################################
-# Set the C++ Standard to be used (Default to C++11)
-####################################################
-set(CXX_STANDARD 14)
-set(CXX_STANDARD_REQUIRED ON)
-set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
-
-####################################################################
-# External libraries
-####################################################################
-find_package(Threads REQUIRED)
-
-####################################################################
-# Set Compiler flags
-####################################################################
-IF (CMAKE_CXX_COMPILER_ID MATCHES "GNU" OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
-  SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Werror -Wpedantic -std=c++${CXX_STANDARD}")
-ELSEIF (MSVC)
-  SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /W4")
-ENDIF()
-
-include_directories("${CMAKE_SOURCE_DIR}/../include")
-
-####################################################
-# Include relevant subdirectories
-####################################################
-# at some point...
-
-####################################################
-# Add Integration Tests file names (without .cpp extension)
-####################################################
-list(APPEND INT_TESTS
-    main
-)
+# Integration tests use private include files.
+include_directories("${CMAKE_SOURCE_DIR}/src")
 
 ####################################################
 # Create executable for each test
 ####################################################
-foreach(TEST ${INT_TESTS})
-    add_executable(int_${TEST} "${CMAKE_SOURCE_DIR}/${TEST}.cpp")
-    target_link_libraries(int_${TEST} Threads::Threads)
-    # At some point this will depend on and link with a library that has the solution (production code)
-    add_test(int_${TEST} int_${TEST})
-endforeach()
-
-add_custom_target(check COMMAND ${CMAKE_CTEST_COMMAND} --verbose)
+add_executable(int_main main.cpp)
+target_link_libraries(int_main Threads::Threads)
+target_link_libraries(int_main solution)
+add_test(int_main int_main)
