+++ tdd-samples/callback-infra/int/main.cpp
@@ -4,6 +4,8 @@
 
 #include "solution.h"
 
+#include "implementation.h"
+
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
 #include "doctest.h"
