+++ tdd-samples/callback-infra/src/CMakeLists.txt
@@ -1,8 +1,8 @@
 
-# add_library(solution STATIC solution.cpp)
+add_library(solution STATIC solution.cpp)
 
 
-# Some boilerplate CMakeLists.txt 
+# Some boilerplate CMakeLists.txt
 # add_executable(exe-name test-file.cpp)
 # target_link_libraries(exe-name PRIVATE solution)
 # add_test(test-name exe-name)
