+++ tdd-samples/callback-infra/src/solution.cpp
@@ -1,0 +1,2 @@
+#include "solution.h"
+#include "implementation.h"
